[IMP] mail: move activity model to common
